Generalise `KiaiFlashingSprite` for arbitrary drawables
