+++ osu.Game.Rulesets.Osu/Skinning/Legacy/KiaiFlashingDrawable.cs
@@ -1,0 +1,50 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System;
+using osu.Framework.Audio.Track;
+using osu.Framework.Graphics;
+using osu.Game.Beatmaps.ControlPoints;
+using osu.Game.Graphics.Containers;
+
+namespace osu.Game.Rulesets.Osu.Skinning.Legacy
+{
+    internal class KiaiFlashingDrawable : BeatSyncedContainer
+    {
+        private readonly Drawable flashingDrawable;
+
+        private const float flash_opacity = 0.3f;
+
+        public KiaiFlashingDrawable(Func<Drawable> creationFunc)
+        {
+            AutoSizeAxes = Axes.Both;
+
+            Children = new[]
+            {
+                creationFunc.Invoke().With(d =>
+                {
+                    d.Anchor = Anchor.Centre;
+                    d.Origin = Anchor.Centre;
+                }),
+                flashingDrawable = creationFunc.Invoke().With(d =>
+                {
+                    d.Anchor = Anchor.Centre;
+                    d.Origin = Anchor.Centre;
+                    d.Alpha = 0;
+                    d.Blending = BlendingParameters.Additive;
+                })
+            };
+        }
+
+        protected override void OnNewBeat(int beatIndex, TimingControlPoint timingPoint, EffectControlPoint effectPoint, ChannelAmplitudes amplitudes)
+        {
+            if (!effectPoint.KiaiMode)
+                return;
+
+            flashingDrawable
+                .FadeTo(flash_opacity)
+                .Then()
+                .FadeOut(timingPoint.BeatLength * 0.75f);
+        }
+    }
+}
